LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/url-resolution-expected.txt

CONSOLE MESSAGE: line 2: SyntaxError: The given initial value does not parse for the given syntax.
CONSOLE MESSAGE: line 2: SyntaxError: The given initial value does not parse for the given syntax.

Harness Error (FAIL), message = SyntaxError: The given initial value does not parse for the given syntax.
